+freebsd 14.x, gcc15, c++23, DevIL 1.8.0
- Extern update
-- boost 1.73.0 -> 1.89.0
-- DevIL 1.7.8 -> 1.8.0
-- cryptopp 8.2.0 -> cryptopp 8.9.0
- c++11 -> c++23
- gcc9 -> gcc15
- Error and warn fix

+++ Srcs/Server/common/item_length.h
@@ -151,7 +151,7 @@ enum EDragonSoulStepTypes
 
 enum EDSInventoryMaxNum
 {
-	DRAGON_SOUL_INVENTORY_MAX_NUM = DS_SLOT_MAX * DRAGON_SOUL_GRADE_MAX * DRAGON_SOUL_BOX_SIZE,
+	DRAGON_SOUL_INVENTORY_MAX_NUM = static_cast<int>(DS_SLOT_MAX) * static_cast<int>(DRAGON_SOUL_GRADE_MAX) * DRAGON_SOUL_BOX_SIZE,
 };
 
 enum EFishSubTypes
